python/requirements.txt
python-dotenv
flask>=1.0.0
flask_restful
eve>=0.8
pymongo==3.10.1
pybase62
